processapp/requirements.txt
click==7.1.2
dnspython==1.16.0
dominate==2.6.0
eventlet==0.29.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paho-mqtt==1.5.1
python-engineio==3.14.2
python-socketio==4.6.1
six==1.15.0
visitor==0.1.3
Werkzeug==1.0.1
boto3==1.16.28
